refactor(bot): remove Slack rollout helpers
