+++ apps/web/src/app/slack/webhook/route.ts
@@ -1,10 +1,8 @@
 import type { NextRequest } from 'next/server';
 import { NextResponse } from 'next/server';
 import { cloneRequestWithBody, handleBotWebhookRequest } from '@/lib/bot/webhook-handler';
-import {
-  ensureSlackIntegrationSyncedForNewBotInfra,
-  getSlackTeamIdFromEventsApiBody,
-} from '@/lib/bot/slack-rollout';
+import { ensureSlackInstallationSyncedForTeam } from '@/lib/bot/slack-installation-sync';
+import { getSlackTeamIdFromEventsApiBody } from '@/lib/slack/request-payload';
 import { verifySlackRequest } from '@/lib/slack/verify-request';
 
 export const maxDuration = 800;
@@ -30,7 +28,7 @@ export async function POST(request: NextRequest) {
 
   try {
     const teamId = getSlackTeamIdFromEventsApiBody(body);
-    await ensureSlackIntegrationSyncedForNewBotInfra(teamId);
+    await ensureSlackInstallationSyncedForTeam(teamId);
   } catch (error) {
     const errorMessage = error instanceof Error ? error.message : String(error);
     console.error('[SlackBot:Webhook] Failed to sync Slack installation before forwarding', {
